der: rename `decodable`/`encodable` modules => `decode`/`encode`
Followup to #523, which renamed the traits but not the modules they're
defined in.

Additionally factors the `DecodeValue`/`EncodeValue` traits into these
modules (originally defined in the `value` module, which is removed in
this PR).
